fix: fix delete all archived chats function

+++ src/archives/archives.service.ts
@@ -74,11 +74,24 @@ export class ArchivesService {
   async deleteAllArchivedChats(userId: string): Promise<void> {
     if (!userId) throw new Error('User Id cannot be empty');
 
-    await this.prisma.chat.deleteMany({
-      where: {
-        userId,
+    const chats = await this.prisma.chat.findMany({
+      where: { 
+        userId ,
         isArchived: true,
       },
     });
+    if (!chats) throw new Error('Chats not found or empty');
+
+    await this.prisma.$transaction([
+      this.prisma.message.deleteMany({
+        where: { chatId: { in: chats.map((chat) => chat.id) } },
+      }),
+      this.prisma.chat.deleteMany({
+        where: { 
+          userId ,
+          isArchived: true,
+        },
+      }),
+    ]);
   }
 }
